Updated Main Models
See DBMS Schema

ToDo:
- Saved Places and Routes
- Aliases

+++ backend/src/main/java/com/orbital/nusmaps/repository/FloorplanRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.orbital.nusmaps.model.Floorplan;
 
-public interface FloorplanRepository extends JpaRepository<Floorplan, Integer> {
+public interface FloorplanRepository extends JpaRepository<Floorplan, Long> {
 }
