+++ backend/src/main/java/com/orbital/nusmaps/service/JwtService.java
@@ -51,8 +51,8 @@ public class JwtService {
         return extractClaim(token, Claims::getSubject);
     }
 
-    public Integer extractUserId(String token) {
-        return extractAllClaims(token).get("userId", Integer.class);
+    public Long extractUserId(String token) {
+        return extractAllClaims(token).get("userId", Long.class);
     }
 
     public boolean extractIsAdmin(String token) {
